refactor: simplify imports with index in Validation
