+++ src/validation/protocols/index.ts
@@ -1,0 +1,2 @@
+export * from './email-validator'
+export * from './username-validator'
