+++ src/validation/validators/compare-fields-validation.ts
@@ -1,5 +1,5 @@
-import { InvalidParamError } from '@/presentation/errors/invalid-param-error'
-import { type Validation } from '@/presentation/protocols/validation'
+import { type Validation } from '@/presentation/protocols'
+import { InvalidParamError } from '@/presentation/errors'
 
 export class CompareFieldsValidation implements Validation {
   constructor(
