+++ src/validation/validators/username-validation.ts
@@ -1,6 +1,6 @@
-import { InvalidParamError } from '@/presentation/errors/invalid-param-error'
-import { type Validation } from '@/presentation/protocols/validation'
-import { type UsernameValidator } from '@/validation/protocols/username-validator'
+import { type UsernameValidator } from '@/validation/protocols'
+import { type Validation } from '@/presentation/protocols'
+import { InvalidParamError } from '@/presentation/errors'
 
 export class UsernameValidation implements Validation {
   constructor(
