+++ src/validation/validators/validation-composite.ts
@@ -1,4 +1,4 @@
-import { type Validation } from '@/presentation/protocols/validation'
+import { type Validation } from '@/presentation/protocols'
 
 export class ValidationComposite implements Validation {
   constructor(private readonly validations: Validation[]) {}
